requirements.txt
asyncio
aiohttp
requests
peewee
peewee_async
aiopg
PyJWT